migrate TWO retrieval to utils
